fix(NS): `NS_GET_PLAYER_POINTS` uintptr_t cast for 64-bit
The macro:

    *(int*)(((int)(player)->pvPrivateData) + NS_PLAYER_POINTS_OFFSET)

casts the pvPrivateData pointer to int before adding the byte offset.
On 64-bit hosts that truncates the upper 32 bits of the pointer
before the addition, so the resulting address is garbage. Same
pattern as the bot.cpp / megahal.cpp / util.h fixes from
`Fix 64-bit pointer truncation and engine-table cast losses`.

Switch to uintptr_t so the arithmetic happens at native pointer
width. No-op on i386 (uintptr_t == int there) and prevents
truncation on amd64 / aarch64. The cast back to (int*) is fine
because the resulting address is in the same private-data region
that pvPrivateData points to.

+++ dlls/bot_const.h
@@ -145,7 +145,7 @@ constexpr int NS_PLAYER_POINTS_OFFSET = NS_PLAYER_EXPERIENCE_OFFSET + 2;
 #endif
 
 //#define NS_GET_PLAYER_POINTS(player) (int)((char*)(player)+NS_PLAYER_POINTS_OFFSET)
-#define NS_GET_PLAYER_POINTS(player) *(int*)(((int)(player)->pvPrivateData)+NS_PLAYER_POINTS_OFFSET)
+#define NS_GET_PLAYER_POINTS(player) *(int*)(((uintptr_t)(player)->pvPrivateData)+NS_PLAYER_POINTS_OFFSET)
 
 /////////////
 // SVENCOOP MONSTERS
